+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Changelog
 
+## v0.0.3
+- Module currently offers:
+  - Displays actions
+  - Clicking on an item action triggers a Roll, Activates an Activity or posts item to chat
+  - Right-click on an item action opens its Sheet  
+  - Clicking on Effects and Conditions toggles them
+  - Right-click on Effects deletes them
+  - Shows Proficiency, Activity and Consume Quantity icons
+  - Shows Roll modifier, Uses and Quantity on appropriate actions
+- Includes the following tabs and actions
+  - Checks & Saves Tab:
+    - Ability Checks
+    - Ability Saves
+    - Skill Checks
+    - Tool Checks
+    - Vehicle Checks
+  - Features Tab:
+    - Class Features (with subgroups)
+    - Talents
+    - Lineage Traits
+    - Heritage Traits
+  - Effects Tab:
+    - Temporary Effects
+    - Passive Effects
+    - Inactive Effects
+    - Conditions
+  - Inventory Tab
+    - Weapons
+    - Ammunition
+    - Armor
+    - Gear (with subgroups)
+    - Tools
+    - Consumables
+    - Sundries
+    - Containers (dynamic groups)
+  - Utility Tab
+    - Combat Actions: [Roll for Initiative], [End Turn]
+    - Token Actions: [Add/Remove to/from Combat], [Make Invisible/Visible], [Toggle Disposition]
+    - Character Actions: Rest Actions
+
 ## v0.0.2
 - Module currently offers:
   - Ability Checks